+++ fdstools/lib/library.py
@@ -178,7 +178,7 @@ def parse_library(handle):
                 value = PAT_SPLIT.split(value)
                 if len(value) != 2:
                     raise ValueError(
-                        "For marker %s, %i flanking sequences were given, need exactly 2 "
+                        "For marker %s, %i flanking sequences were given, need exactly two "
                         "(note: specify 'REF' to automatically load genome reference)"
                         % (marker, len(value)))
                 for i, val in enumerate(value):
@@ -186,15 +186,29 @@ def parse_library(handle):
                         value[i] = ""
                     elif val.isdigit():
                         raise ValueError(
-                            "Specifying a flank length (%s for marker %s) in the "
-                            "library file is no longer supported; specify 'REF' "
-                            "to automatically load genome reference"
+                            "Specifying a flank length (%s for marker %s) in the [flanks] "
+                            "section of the library file is no longer supported. "
+                            "Use the [flank_length] section to specify marker-specific flank "
+                            "length limitations in the library file; specify 'REF' in the "
+                            "[flanks] section to automatically load genome reference"
                             % (val, marker))
                     elif PAT_SEQ_IUPAC.match(val) is None:
                         raise ValueError(
                             "Flanking sequence '%s' of marker %s is not a valid sequence "
                             "(note: specify 'REF' to automatically load genome reference)"
                             % (val, marker))
+            elif section_low == "flank_length":
+                value = PAT_SPLIT.split(value)
+                if len(value) != 4:
+                    raise ValueError(
+                        "Flank length should be four numbers, but %i values were "
+                        "given for marker %s" % (len(value), marker))
+                for i, val in enumerate(value):
+                    if not val.isdigit() or not int(val):
+                        raise ValueError(
+                            "Flank length '%s' of marker %s is not a valid positive "
+                            "integer" % (val, marker))
+                    value[i] = int(val)
             elif section_low in ("prefix", "suffix"):
                 if not PAT_SEQ_RAW.match(value):
                     raise ValueError(