+++ fdstools/tools/library.py
@@ -103,10 +103,9 @@ def make_empty_library_ini(type, microhaplotypes=False):
     ini.add_comment("flanks",
         "The TSSV tool will use a pair of short anchor sequences just outside the reported range "
         "of each marker (e.g., primer sequences) to identify which input reads correspond to "
-        "which marker. The number of flanking bases used by TSSV can be controlled with the "
-        "-L/--flank-length option of that tool. %s The sequence may contain IUPAC codes for "
-        "ambiguous positions to account for degenerate bases in the primers or for bisulfite-"
-        "converted targets in methylation-based studies (e.g., Y matches either C or T)." % (
+        "which marker. %s The sequence may contain IUPAC codes for ambiguous positions to account "
+        "for degenerate bases in the primers or for bisulfite-converted targets in methylation-"
+        "based studies (e.g., Y matches either C or T)." % (
             "Specify two comma-separated values: left flank and right flank sequence, in the same "
             "sequence orientation (strand)."
                 if type in ("str", "non-str") else
@@ -116,6 +115,13 @@ def make_empty_library_ini(type, microhaplotypes=False):
             "explicit anchor sequence to use.%s" % (
                 " For markers configured explicitly in this library file, the anchor sequences "
                 "MUST be specified explicitly as well." if type == "full" else ""))))
+    ini.add_section("flank_length")
+    ini.add_comment("flank_length",
+        "The number of flanking bases used by TSSV can be controlled with the -L/--flank-length "
+        "option of that tool. If the flank length of a specific marker needs to be limited to a "
+        "specific value or range of values, this can be specified here. Provide four numbers "
+        "for each such marker: the minimum and maximum possible length for the left flank, "
+        "followed by the minimum and maximum possible length for the right flank.")
     ini.add_section("max_expected_copies")
     ini.add_comment("max_expected_copies",
         "By default, the Allelefinder tool will report up to 2 alleles per marker, but only a "